extract index manager and switcher to separate files; add empty patch fetcher;
